requirements.txt
python-telegram-bot
google-generativeai
pillow
pymongo
python-dotenv
emoji